Update decoders to support fallback ErrorResponse
Update FallbackReaderStrategy to decode any response into an
ErrorResponse if that type is requested while reading. Also include
it to the list of default reader strategies (decoders) in our
DiscordWebClient.

This is needed to decode text/html CloudFlare responses. After this
commit they will appear in the ErrorResponse field of a resulting
ClientException on error status codes.

For normal status codes, the response body will be output as String
if no other decoder is able to process it before.