Started README Overhaul
READMEs are important for open-source projects. They provide a first impression for their projects. They should tell you what it is, how to get it, and how to use it. It should be inviting, easy to follow, and also nice to look at. I plan on revamping the README to include a *lot* more examples and to just flesh out ways to make ourselves stand out and provide the essential question, *"Why should I use Discord4J?"

This isn't finished yet. But feedback is welcome.