Modify RestClientBuilder to be more extensible and apply changes to token and auth scheme.
The builder itself has been added as a generic parameter, B, and references to 'this' have been replaced to references of B such that subclasses can chain methods without losing the type reference of the subclass. The configuration of the builder has been renamed to Resources and added as a generic parameter R so that subclasses can use custom builder resources that are applied when the client is built. Boilerplate code has also been simplified.